python/terminal-mcp-orchestrator/requirements.txt
# Core dependencies
requests>=2.28.0
sseclient-py>=1.7.2

# Development dependencies (optional)
pytest>=7.0.0 